Added update, delete and search functionality

+++ main.py
@@ -10,6 +10,9 @@ class MainApp(MDApp):
         # set the theme color
         self.theme_cls.primary_palette = "DeepPurple"
 
+        # row id to keep track of the id that have been selected
+        self.row_ids = []
+
         #create the database and connect
         self.database = Database()
 
@@ -27,19 +30,22 @@ class MainApp(MDApp):
             ],
             row_data = [],
         )
-
+        self.data_table.bind(on_check_press=self.on_check_press)
         self.root.ids['datatable'].add_widget(self.data_table)
-        self.data_table.row_data = [
-            ('1', 'Food for thought', 21.56, '21/04/2021'),
-            ('2', 'Food for thought', 21.56, '21/04/2021'),
-            ('3', 'Food for thought', 21.56, '21/04/2021'),
-            ('4', 'Food for thought', 21.56, '21/04/2021'),
-            ('5', 'Food for thought', 21.56, '21/04/2021'),
-            ('6', 'Food for thought', 21.56, '21/04/2021'),
-            ('7', 'Food for thought', 21.56, '21/04/2021')
-        ]
 
-   # Data table
+    # data table
+    def on_check_press(self, instance_table, current_row):
+        '''Called when the checkbox in the data table is clicked'''
+        print(current_row)
+        self.row_ids.append(current_row[0])
+        self.root.ids['description_field'].text = current_row[1]
+        self.root.ids['price_field'].text = current_row[2]
+        self.root.ids['date_field'].text = current_row[3]
+
+   # Date picker
+    def on_start(self):
+       '''Called when the program is started'''
+       self.reload_data_table()
 
     def on_save(self, instance, value, date_range):
         '''Called when we select a date from the datetime picker'''
@@ -58,11 +64,15 @@ class MainApp(MDApp):
     # Search
     def search(self):
         '''Search values in the database'''
-        print(self.root.ids['search'].text)
+        id_check = self.root.ids['id_check']
 
-    def insert_date(self, description, price, date):
-        '''Inserts data into the database'''
-        self.database.create_entry(description, price, date)
+        if id_check.active == True:
+            try:
+                self.data_table.row_data = self.database.get_product_by_id(int(self.root.ids['search'].text))
+            except:
+                pass
+        else:
+            self.data_table.row_data = self.database.get_products_by_description(str(self.root.ids['search'].text))
 
     
     # Database Operations
@@ -74,20 +84,44 @@ class MainApp(MDApp):
 
         self.database.create_entry(description, price, date)
         self.reload_data_table()
-        
-    def update_data(self, id,description, price, date):
-        '''Update the data in the database for the given primary key'''
-        self.database.update_product(id, description, price, date)
-        self.reload_data_table()
+        self.clear_text_entries()
 
-    def delete_data(self, id):
+        
+    def update_data(self):
+        '''Update the data in the database for the given primary key'''
+        description = self.root.ids['description_field'].text
+        price = self.root.ids['price_field'].text
+        date = self.root.ids['date_field'].text
+        if self.row_ids != []:
+            self.database.update_product(self.row_ids[-1], description, price, date)
+            self.reload_data_table()
+        
+        self.row_ids = []
+        self.clear_text_entries()
+
+
+    def delete_data(self):
         '''Delete data with the given id'''
-        self.database.delete_product(id)
-        self.reload_data_table()
+        if self.row_ids != []:
+            for id in self.row_ids:
+                self.database.delete_product(id)
+
+            self.reload_data_table()
+            self.row_ids = []
+            self.clear_text_entries()   
+
+        else:
+            print("Select rows first!")
 
     def reload_data_table(self):
         '''Reload the datatable after an action e.g update, delete or insert'''
         self.data_table.row_data = self.database.get_products()
+
+    def clear_text_entries(self):
+        '''Clears the text entries after an action'''
+        self.root.ids['description_field'].text = ''
+        self.root.ids['price_field'].text = ''
+        self.root.ids['date_field'].text = ''
